photo upload update and delete on staff

+++ controllers/admin_controller.js
@@ -57,7 +57,7 @@ router.get("/products/:id/edit", (req, res) => {
 });
 
 //update - update database
-router.put("/products/:id",upload.single("photo"), (req, res) => {
+router.put("/products/:id", upload.single("photo"), (req, res) => {
     db.Product.update(req.body, {
         where: {
             id: req.params.id
@@ -69,8 +69,11 @@ router.put("/products/:id",upload.single("photo"), (req, res) => {
 
 //delete product in database
 router.delete("/products/:id/:name", (req, res) => {
+    //Check if file exist and delete
     var filePath = "./public/assets/images/productUpload/" + req.params.name + ".png";
-    fs.unlinkSync(filePath); 
+    if(fs.existsSync(filePath)){
+        fs.unlinkSync(filePath);
+    }
     db.Product.destroy({
         where: {
             id: req.params.id
@@ -98,7 +101,7 @@ storage = multer.diskStorage({
     }
 });
 upload = multer({ storage: storage });
-router.post("/services/new", upload.single("photo"),(req, res) => {
+router.post("/services/new", upload.single("photo"), (req, res) => {
     // console.log(req.body)
     db.Service.create(
         {
@@ -143,8 +146,12 @@ router.put("/services/:id", upload.single("photo"), (req, res) => {
 
 //delete service in database
 router.delete("/services/:id/:name", (req, res) => {
+    //Check if file exist and delete
     var filePath = "./public/assets/images/serviceUpload/" + req.params.name + ".png";
-    fs.unlinkSync(filePath); 
+        if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+        }
+   
     db.Service.destroy({
         where: {
             id: req.params.id
@@ -225,7 +232,14 @@ router.get("/staff", (req, res) => {
 });
 
 //add staff in database
-router.post("/staff/new", (req, res) => {
+storage = multer.diskStorage({
+    destination: "./public/assets/images/staffUpload",
+    filename: function (req, file, cb) {
+        cb(null, req.body.name + ".png");
+    }
+});
+upload = multer({ storage: storage });
+router.post("/staff/new", upload.single("photo"), (req, res) => {
     var emailId;
     var addressId;
     var phoneId;
@@ -264,6 +278,7 @@ router.post("/staff/new", (req, res) => {
             hour: req.body.hour,
             emergency_contact_name: req.body.emergency_contact_name,
             emergency_contact_phone: req.body.emergency_contact_phone,
+            photo: req.body.name + ".png",
             comment: req.body.comment,
             EmailId: emailId,
             AddressId: addressId,
@@ -298,7 +313,7 @@ router.get("/staff/:id/edit", (req, res) => {
 });
 
 //update - update database
-router.put("/staff/:id", (req, res) => {
+router.put("/staff/:id", upload.single("photo"), (req, res) => {
     db.Email.update({
         email: req.body.email
     }, {
@@ -331,7 +346,7 @@ router.put("/staff/:id", (req, res) => {
                     hour: req.body.hour,
                     emergency_contact_name: req.body.emergency_contact_name,
                     emergency_contact_phone: req.body.emergency_contact_phone,
-                    photo: req.body.photo,
+                    photo: req.body.name + ".png",
                     comment: req.body.comment
                 }, {
                         where: {
@@ -343,8 +358,12 @@ router.put("/staff/:id", (req, res) => {
 });
 
 //delete staff in database
-router.delete("/staff/:id", (req, res) => {
-    // console.log(req.params.id);
+router.delete("/staff/:id/:name", (req, res) => {
+    //Check if file exist and delete
+    var filePath = "./public/assets/images/staffUpload/" + req.params.name + ".png";
+    if (fs.existsSync(filePath)) {
+        fs.unlinkSync(filePath);
+    }
     db.Staff.destroy({
         where: {
             id: req.params.id
@@ -564,7 +583,7 @@ router.put("/membership/:id", (req, res) => {
     }).then(data => {
         res.redirect("/admin/membership")
     }
-        )
+    )
 });
 
 //delete membership in database
